refactor: Style 모듈의 폴더 구조 변경
