+++ ui/accordion/component.tsx
@@ -3,7 +3,7 @@
 import { forwardRef } from "react";
 import * as Primitive from "@radix-ui/react-accordion";
 
-import { tv, VariantProps } from "@/utils/tailwind-variants";
+import { tv, VariantProps } from "@/utils/style/tailwind-variants";
 
 export type AccordionRootProps =
   | Primitive.AccordionSingleProps
