+++ ui/check-box.tsx
@@ -2,7 +2,7 @@
 
 import * as Primitive from "@radix-ui/react-checkbox";
 
-import { tv, VariantProps } from "@/utils/tailwind-variants";
+import { tv, VariantProps } from "@/utils/style/tailwind-variants";
 import CheckIcon from "./icons/check";
 
 export const checkBoxVariant = tv({
