+++ ui/dialog/component.tsx
@@ -2,8 +2,8 @@
 
 import * as Primitive from "@radix-ui/react-dialog";
 
-import { tv, VariantProps } from "@/utils/tailwind-variants";
-import { sem } from "@/utils/semantic-styles";
+import { tv, VariantProps } from "@/utils/style/tailwind-variants";
+import { sem } from "@/utils/style/semantic-styles";
 import CloseIcon from "../icons/close";
 
 export const dialogVariant = tv({
