+++ ui/settings/tab-button.tsx
@@ -2,8 +2,8 @@
 
 import { ComponentPropsWithoutRef, forwardRef } from "react";
 
-import { tv, VariantProps } from "@/utils/tailwind-variants";
-import { sem } from "@/utils/semantic-styles";
+import { tv, VariantProps } from "@/utils/style/tailwind-variants";
+import { sem } from "@/utils/style/semantic-styles";
 
 export const tabButtonVariant = tv({
   base: sem()
