+++ ui/slider.tsx
@@ -2,7 +2,7 @@
 
 import * as Primitive from "@radix-ui/react-slider";
 
-import { tv, VariantProps } from "@/utils/tailwind-variants";
+import { tv, VariantProps } from "@/utils/style/tailwind-variants";
 
 export const sliderVariant = tv({
   slots: {
